+++ tests/settings_swapped.py
@@ -3,4 +3,5 @@ from .settings import *  # noqa
 
 OAUTH2_PROVIDER_ACCESS_TOKEN_MODEL = "tests.SampleAccessToken"
 OAUTH2_PROVIDER_APPLICATION_MODEL = "tests.SampleApplication"
+OAUTH2_PROVIDER_DEVICE_GRANT_MODEL = "tests.SampleDeviceGrant"
 OAUTH2_PROVIDER_REFRESH_TOKEN_MODEL = "tests.SampleRefreshToken"