+++ CHANGELOG.md
@@ -13,7 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 <!-- ### Fixed -->
 <!-- ### Security -->
 
-## [unreleased]
+## [3.3.0] - 2025-05-21
 
 ### Added
 * #1637 Support for Django 6.0